refactor: move system service to system info repositiry service
