+++ src/extension/features/system/thunks/saveDisableWhatsNewOnUpdateThunk.ts
@@ -3,12 +3,14 @@ import { AsyncThunk, createAsyncThunk } from '@reduxjs/toolkit';
 // enums
 import { ThunkEnum } from '../enums';
 
-// services
-import SystemService from '@extension/services/SystemService';
+// errors
+import { MalformedDataError } from '@extension/errors';
+
+// repositories
+import SystemInfoRepositoryService from '@extension/repositories/SystemInfoRepositoryService';
 
 // types
 import type { IBaseAsyncThunkConfig, IMainRootState } from '@extension/types';
-import { MalformedDataError } from '@extension/errors';
 
 const saveDisableWhatsNewOnUpdateThunk: AsyncThunk<
   boolean, // return
@@ -29,9 +31,7 @@ const saveDisableWhatsNewOnUpdateThunk: AsyncThunk<
       return rejectWithValue(new MalformedDataError(_error));
     }
 
-    const { whatsNewInfo } = await new SystemService({
-      logger,
-    }).saveToStorage({
+    const { whatsNewInfo } = await new SystemInfoRepositoryService().save({
       ...systemInfo,
       whatsNewInfo: {
         ...systemInfo.whatsNewInfo,
