--- src/extension/services/SystemService/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SystemService';
-export * from './types';
