--- src/extension/services/SystemService/types/index.ts
@@ -1,1 +1,0 @@
-export type { default as ICreateOptions } from './ICreateOptions';
